fix: unblock the system temp directory on macOS, and pin the lint rules
Two CI failures, both real.

macOS refused every path under the system temp directory. /tmp and /var are
symlinks into /private there, so realpath turns every temporary directory into
/private/var/folders/..., which sits inside two of the POSIX system roots added
for Layer 1. Fourteen tests failed while Linux and Windows were green, and the
same bug made scripts/make_sandbox.py useless on macOS: the demo tree it builds
in the temp directory could never be scanned.

POSIX_TEMP_ROOTS exempts that scratch space from the system rules only, and is
deliberately narrow - the directories the OS itself empties, nothing else.
/private/etc, /private/var/db, /private/var/root and /var/log stay blocked, and
a data-level test asserts that on every platform so the carve-out cannot widen
unnoticed. Minimum depth, profile rules, protected names, user-protected paths
and reparse points all still apply inside temp.

Lint failed because ruff was running with its built-in defaults, which drift
between releases. The rule set is now pinned in pyproject, with each ignore
carrying the reason it exists: Typer puts calls in argument defaults, the
fail-closed paths use try/except/pass on purpose, and Windows environment
variables really are mixed case.

Ruff also found three dead 'caps' locals left over from the MascotBox to
MascotLine refactor, and make_sandbox.py fell back to '.' when %TEMP% was
absent, which would have built the sandbox inside the repository on Linux and
macOS.

+++ dtcleaner/utils/formatting.py
@@ -82,5 +82,5 @@ def bar(part: float, total: float, width: int = 20, fill: str = "█", empty: st
     if total <= 0:
         return empty * width
     ratio = max(0.0, min(1.0, part / total))
-    filled = int(round(ratio * width))
+    filled = round(ratio * width)
     return fill * filled + empty * (width - filled)
